Add a generic navigation targets provider.
This can be used to get the default menu. It also accepts an optional
viewRefCollection which can be used to provides the ElementRef
associated with each default navigation target.
