+++ src/app/blog/pages/blog-page/blog-page.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { BlogPost, BlogPostRepository } from 'src/app/core/models/blog-post.models';
-import { NavbarItem } from 'src/app/ui/navbar/navbar.component';
+import { NavbarItem, getDefaultNavigationTarget } from 'src/app/ui/shared/models/navbar.models';
 import { BLOG_POST_REPOSITORY } from 'src/app/core/config/injection-tokens';
 
 @Component({
@@ -14,24 +14,7 @@ export class BlogPageComponent implements OnInit {
   public post: BlogPost;
 
   public get navigationTargets(): NavbarItem[] {
-    return [
-      {
-        label: 'Accueil',
-        navigationTarget: { routerLink: '/home', fragment: 'hero' }
-      },
-      {
-        label: 'À Propos',
-        navigationTarget: { routerLink: '/home', fragment: 'about' }
-      },
-      {
-        label: 'Compétences',
-        navigationTarget: { routerLink: '/home', fragment: 'skills' }
-      },
-      {
-        label: 'Publications',
-        navigationTarget: { routerLink: '/home', fragment: 'blog-posts' }
-      }
-    ]
+    return getDefaultNavigationTarget();
   }
 
   constructor(
@@ -46,5 +29,4 @@ export class BlogPageComponent implements OnInit {
     if( this.post === undefined )
       throw new Error(`Unable to load post associated with path: ${lastSegment.path}`);
   }
-
 }
