+++ src/app/home/pages/home-page/home-page.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
-import { NavbarItem } from 'src/app/ui/navbar/navbar.component';
 import { HeroAreaComponent } from '../../components/hero-area/hero-area.component';
 import { AboutSectionComponent } from '../../components/about-section/about-section.component';
 import { SkillsSectionComponent } from '../../components/skills-section/skills-section.component';
 import { BlogSectionComponent } from '../../components/blog-section/blog-section.component';
+import { NavbarItem, NavbarItemViewRefCollection, getDefaultNavigationTarget } from 'src/app/ui/shared/models/navbar.models';
 
 @Component({
   selector: 'app-home-page',
@@ -18,28 +18,12 @@ export class HomePageComponent implements OnInit {
   @ViewChild(BlogSectionComponent, { read: ElementRef, static: true }) blogSectionComponentRef: ElementRef;
 
   public get navigationTargets(): NavbarItem[] {
-    return [
-      {
-        label: 'Accueil',
-        navigationTarget: { routerLink: '/home', fragment: 'hero' },
-        viewRef: this.heroAreaComponentRef
-      },
-      {
-        label: 'À Propos',
-        navigationTarget: { routerLink: '/home', fragment: 'about' },
-        viewRef: this.aboutComponentRef
-      },
-      {
-        label: 'Compétences',
-        navigationTarget: { routerLink: '/home', fragment: 'skills' },
-        viewRef: this.skillsComponentRef
-      },
-      {
-        label: 'Publications',
-        navigationTarget: { routerLink: '/home', fragment: 'blog-posts' },
-        viewRef: this.blogSectionComponentRef
-      }
-    ];
+    return getDefaultNavigationTarget({
+      'home': this.heroAreaComponentRef,
+      'about': this.aboutComponentRef,
+      'skills': this.skillsComponentRef,
+      'posts': this.blogSectionComponentRef
+    });
   }
 
   constructor() { }
