+++ src/app/ui/components/navbar/navbar.component.less
@@ -1,4 +1,4 @@
-@import '../../../styles.less';
+@import '../../../../styles.less';
 
 .navbar {
     background-color: #1d2024;
