+++ src/app/ui/components/section-header/section-header.component.less
@@ -1,4 +1,4 @@
-@import '../../../styles.less';
+@import '../../../../styles.less';
 
 .title {
